# PERF: Improve initial paint times