--- a/sass/example.sass
+++ b/sass/example.sass
@@ -100,6 +100,8 @@
     width: 4px
 
 .example-output
+  body.has-js &
+    display: none
   border: 1px solid $border
   border-bottom-left-radius: 4px
   border-bottom-right-radius: 4px