In order to improve initial paint times, I've made a change so that the examples are hidden on first load. They are then shown as the scroll position changes.

On my Mid 2014 MBP in Chrome the results are:

*Before:*
- Rendering: ~3000ms
- Painting: ~350ms

*After:*
- Rendering: ~2000ms
- Painting: ~80ms

---

This will improve the performance of the site as it currently is, but I'd recommend splitting the single page up into multiple, smaller pages. Perhaps split by sections of the css spec? This will have a much greater effect on performance.